Invalid path'ing on CDN backed by Storage Account
We just went through setting up a CDN endpoint per this article. It turns out that the files just need to be in a public container on the storage account, that container has no impact on the final URL. This was revealed to us after banging away for about an hour against a file we could see in the storage account, but CDN was returning a 404. Someone typo'd the URL excluding the public container and the file began to download to their machine. I only edited the portion related to Storage Accounts, as that's all we are doing.